requirements.txt
langgraph>=0.2.0
langchain-core>=0.2.0
boto3>=1.34.0
python-dotenv>=1.0.0
langchain-aws>=0.1.0
fastapi>=0.110.0
uvicorn>=0.22.0
requests>=2.31.0
